Revert "package structure refactor"
This reverts commit 4b6dd3b6938fb9ad2919551f36b82a058f77f00b.
