+++ src/mcp_as_a_judge/prompts/system/judge_coding_plan.md
