+++ src/mcp_as_a_judge/prompts/system/research_validation.md
