+++ src/mcp_as_a_judge/prompts/user/judge_coding_plan.md
