+++ src/mcp_as_a_judge/prompts/user/research_validation.md
