+++ src/prompts/system/dynamic_schema.md
