+++ src/prompts/system/get_workflow_guidance.md
