+++ src/prompts/system/research_validation.md
