+++ src/prompts/user/dynamic_schema.md
