+++ src/prompts/user/get_workflow_guidance.md
