+++ src/prompts/user/research_validation.md
